convert mention plugin to typescript
