+++ draft-js-mention-plugin/package.json
@@ -44,5 +44,8 @@
   "peerDependencies": {
     "draft-js": "^0.10.1 || ^0.11.0",
     "react": "^16.3.0"
+  },
+  "devDependencies": {
+    "@types/lodash": "^4.14.162"
   }
 }
