--- draft-js-mention-plugin/src/MentionSuggestions/Entry/Avatar/Avatar.js
@@ -1,17 +1,0 @@
-import React from 'react';
-
-const Avatar = ({ mention, theme = {} }) => {
-  if (mention.avatar) {
-    return (
-      <img
-        src={mention.avatar}
-        className={theme.mentionSuggestionsEntryAvatar}
-        role="presentation"
-      />
-    );
-  }
-
-  return null;
-};
-
-export default Avatar;
