--- draft-js-mention-plugin/src/MentionSuggestions/Entry/DefaultEntryComponent.js
@@ -1,21 +1,0 @@
-import React from 'react';
-import Avatar from './Avatar/Avatar';
-
-const DefaultEntryComponent = props => {
-  const {
-    mention,
-    theme,
-    isFocused, // eslint-disable-line no-unused-vars
-    searchValue, // eslint-disable-line no-unused-vars
-    ...parentProps
-  } = props;
-
-  return (
-    <div {...parentProps}>
-      <Avatar mention={mention} theme={theme} />
-      <span className={theme.mentionSuggestionsEntryText}>{mention.name}</span>
-    </div>
-  );
-};
-
-export default DefaultEntryComponent;
