--- draft-js-mention-plugin/src/utils/getTypeByTrigger.js
@@ -1,4 +1,0 @@
-const getTypeByTrigger = trigger =>
-  trigger === '@' ? 'mention' : `${trigger}mention`;
-
-export default getTypeByTrigger;
