+++ draft-js-mention-plugin/src/utils/getTypeByTrigger.ts
@@ -1,0 +1,3 @@
+export default function getTypeByTrigger(trigger: string): string {
+  return trigger === '@' ? 'mention' : `${trigger}mention`;
+}
