+++ package.json
@@ -33,6 +33,7 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^10.4.9",
     "@testing-library/react-hooks": "^3.4.1",
+    "@types/jest": "^26.0.15",
     "@typescript-eslint/eslint-plugin": "^4.1.0",
     "alex": "^4.0.1",
     "animateplus": "^1.4.0",
@@ -65,7 +66,7 @@
     "extract-text-webpack-plugin": "^2.0.0",
     "flow-bin": "^0.69.0",
     "husky": "^3.0.4",
-    "jest": "^26.4.2",
+    "jest": "^26.6.0",
     "jsdom": "^11.3.0",
     "linaria": "^1.4.0-alpha.1",
     "lint-staged": "^9.2.3",
@@ -90,7 +91,9 @@
     "sinon-chai": "^2.8.0",
     "style-loader": "^0.13.2",
     "superagent": "^3.5.0",
-    "typescript": "4.0.2",
+    "ts-jest": "^26.4.1",
+    "ts-node": "^9.0.0",
+    "typescript": "^4.0.3",
     "webpack": "^2.2.1",
     "webpack-dev-middleware": "^1.10.1",
     "webpack-hot-middleware": "^2.17.1"
@@ -136,7 +139,7 @@
     "lint:eslint:fix": "eslint --fix --rule 'mocha/no-exclusive-tests: 2' ./",
     "lint:flow": "flow status",
     "storybook": "start-storybook -p 6006 -s ./stories/public",
-    "test": "mocha --compilers @babel/register --exclude \"./*/src/**/__test__/*test.js\" --require testHelper.js \"./*/src/**/__test__/*.js\"",
+    "test": "env TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\" }' mocha -r ts-node/register --compilers @babel/register --exclude \"./*/src/**/__test__/*test.js\" --require testHelper.js \"./*/src/**/__test__/*.js\"",
     "test:jest": "jest",
     "test:watch": "yarn test -- --watch"
   },
